Fix template inference and types

+++ src/examples/convolution/test_convolution.cc
@@ -32,7 +32,7 @@ void test_convolution(std::size_t N, std::size_t K, Dimension axis, T precision,
   for (int i = 0; i < N; ++i) {
     // T expnt = 1500 + 1500*drand48();
     mra::Coordinate<T,NDIM> r;
-    for (size_t d=0; d<NDIM; d++) {
+    for (int d=0; d<NDIM; d++) {
       r[d] = T(-6.0) + T(12.0)*drand48();
     }
     std::cout << "Gaussian " << i << " expnt " << expnt << std::endl;
